+++ packages/leva/src/components/UI/StyledUI.ts
@@ -4,6 +4,7 @@ import { StyledContent } from '../Folder/StyledFolder'
 
 export const StyledRow = styled('div', {
   position: 'relative',
+  zIndex: 100,
   display: 'grid',
   rowGap: '$leva__rowGap',
   gridTemplateRows: 'minmax($sizes$leva__rowHeight, max-content)',
@@ -49,6 +50,16 @@ export const CopyLabelContainer = styled('div', {
     backgroundColor: '$leva__elevation2',
   },
   '&:hover > div > svg': { display: 'block' },
+
+  variants: {
+    align: {
+      top: {
+        height: '100%',
+        alignItems: 'flex-start',
+        paddingTop: '$leva__sm',
+      },
+    },
+  },
 })
 
 export const StyledOptionalToggle = styled('input', {
@@ -62,7 +73,6 @@ export const StyledOptionalToggle = styled('input', {
     position: 'relative',
     $leva__flexCenter: '',
     height: '100%',
-    WebkitUserSelect: 'none', // TODO remove when stitches fixes prefix import
     userSelect: 'none',
     cursor: 'pointer',
     paddingLeft: 2,
@@ -107,15 +117,6 @@ export const StyledLabel = styled('label', {
   whiteSpace: 'nowrap',
   '& > svg': {
     display: 'block', // fixes svg vertical misalignment
-  },
-  variants: {
-    align: {
-      top: {
-        height: '100%',
-        alignItems: 'flex-start',
-        paddingTop: '$leva__sm',
-      },
-    },
   },
 })