+++ packages/leva/src/types/utils.ts
@@ -8,7 +8,11 @@
  * type B = Id<{ a: 'a' } & { b: 'b' }> // { a: 'a', b: 'b' }
  * ```
  */
-export type BeautifyUnionType<T> = T extends Function ? T : T extends infer TT ? { [k in keyof TT]: TT[k] } : never
+export type BeautifyUnionType<T> = T extends Function
+  ? T
+  : T extends infer TT
+  ? { [k in keyof TT]: TT[k] } & IterableIterator<T> // adds Iterator to the return type in case it has any
+  : never
 
 export type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (k: infer I) => void ? I : never