+++ packages/plugin-bezier/src/index.ts
@@ -1,0 +1,11 @@
+import { createPlugin, formatVector } from 'leva/plugin'
+import { Bezier } from './Bezier'
+import { normalize, sanitize } from './bezier-plugin'
+import { InternalBezierSettings } from './bezier-types'
+
+export const bezier = createPlugin({
+  normalize,
+  sanitize,
+  format: (value: any, settings: InternalBezierSettings) => formatVector(value, settings),
+  component: Bezier,
+})